+++ internal/types/types_test.go
@@ -1,7 +1,0 @@
-package types
-
-import "testing"
-
-func TestTypes(t *testing.T) {
-	// Dummy test
-}